refactored escaping so that Resource.export() also escapes
