+++ import_export/formats/base_formats.py
@@ -1,5 +1,3 @@
-import html
-
 import tablib
 from tablib.formats import registry
 
@@ -85,10 +83,6 @@ class TablibFormat(Format):
         return tablib.import_set(in_stream, format=self.get_title(), **kwargs)
 
     def export_data(self, dataset, **kwargs):
-        if kwargs.pop("escape_html", None):
-            self._escape_html(dataset)
-        if kwargs.pop("escape_formulae", None):
-            self._escape_formulae(dataset)
         return dataset.export(self.get_title(), **kwargs)
 
     def get_extension(self):
@@ -102,21 +96,6 @@ class TablibFormat(Format):
 
     def can_export(self):
         return hasattr(self.get_format(), "export_set")
-
-    def _escape_html(self, dataset):
-        for _ in dataset:
-            row = dataset.lpop()
-            row = [html.escape(str(cell)) for cell in row]
-            dataset.append(row)
-
-    def _escape_formulae(self, dataset):
-        def _do_escape(s):
-            return s.replace("=", "", 1) if s.startswith("=") else s
-
-        for _ in dataset:
-            row = dataset.lpop()
-            row = [_do_escape(str(cell)) for cell in row]
-            dataset.append(row)
 
 
 class TextFormat(TablibFormat):
