+++ import_export/resources.py
@@ -1095,14 +1095,26 @@ class Resource(metaclass=DeclarativeMetaclass):
             queryset = self.get_queryset()
         queryset = self.filter_export(queryset, *args, **kwargs)
         headers = self.get_export_headers()
-        data = tablib.Dataset(headers=headers)
+        dataset = tablib.Dataset(headers=headers)
 
         for obj in self.iter_queryset(queryset):
-            data.append(self.export_resource(obj))
+            dataset.append(self.export_resource(obj))
 
-        self.after_export(queryset, data, *args, **kwargs)
+        if kwargs.get("escape_formulae") is True:
+            self._escape_formulae(dataset)
 
-        return data
+        self.after_export(queryset, dataset, *args, **kwargs)
+
+        return dataset
+
+    def _escape_formulae(self, dataset):
+        def _do_escape(s):
+            return s.replace("=", "", 1) if s.startswith("=") else s
+
+        for _ in dataset:
+            row = dataset.lpop()
+            row = [_do_escape(str(cell)) for cell in row]
+            dataset.append(row)
 
     def _get_ordered_field_names(self, order_field):
         """
