+++ pyproject.toml
@@ -36,7 +36,7 @@ classifiers = [
 dependencies = [
     "diff-match-patch",
     "Django>=3.2",
-    "tablib[html,ods,xls,xlsx,yaml]==3.5.0",
+    "tablib[html,ods,xls,xlsx,yaml]@git+https://github.com/jazzband/tablib.git@master#egg=tablib"
 ]
 
 [project.urls]
