+++ requirements/base.txt
@@ -1,4 +1,3 @@
 Django>=3.2
-# tablib temporarily pinned to 3.5.0 - see issue #1602
-tablib[html,ods,xls,xlsx,yaml]==3.5.0
+tablib[html,ods,xls,xlsx,yaml]@git+https://github.com/jazzband/tablib.git@master#egg=tablib
 diff-match-patch
