+++ tests/core/tests/test_base_formats.py
@@ -211,31 +211,16 @@ class HTMLFormatTest(TestCase):
         )
 
     def test_export_html_escape(self):
-        res = self.format.export_data(self.dataset, escape_html=True)
-        self.assertIn(
-            (
-                "<tr><td>1</td>\n"
-                "<td>good_user</td>\n"
-                "<td>John Doe</td></tr>\n"
-                "<tr><td>2</td>\n"
-                "<td>evil_user</td>\n"
-                "<td>&lt;script&gt;alert(&quot;I want to steal your credit card data"
-                "&quot;)&lt;/script&gt;</td></tr>\n"
-            ),
-            res,
-        )
-
-    def test_export_data_no_escape(self):
         res = self.format.export_data(self.dataset)
         self.assertIn(
             (
-                "<tr><td>1</td>\n"
-                "<td>good_user</td>\n"
-                "<td>John Doe</td></tr>\n"
-                "<tr><td>2</td>\n"
-                "<td>evil_user</td>\n"
-                '<td><script>alert("I want to steal your credit card data")'
-                "</script></td></tr>\n"
+                "<tr><td>1</td>"
+                "<td>good_user</td>"
+                "<td>John Doe</td></tr>"
+                "<tr><td>2</td>"
+                "<td>evil_user</td>"
+                '<td>&lt;script&gt;alert("I want to steal your credit card data")'
+                "&lt;/script&gt;</td></tr>"
             ),
             res,
         )
